% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASBO-package.R
\docType{package}
\name{ASBO-package}
\alias{ASBO}
\alias{ASBO-package}
\title{ASBO: Anomaly Detection by Options using the PELT Algorithm}
\description{
THIS IS OUT OF DATE Implimentations of the univariate CAPA \doi{10.1002/sam.11586} and PELT \doi{10.1080/01621459.2012.737745} algotithms along with various cost functions for different distributions and models. The modular design, using R6 classes, favour ease of extension (for example user written cost functions) over the performance of other implimentations (e.g. \doi{10.32614/CRAN.package.changepoint}, \doi{10.32614/CRAN.package.anomaly}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/waternumbers/ASBO}
  \item Report bugs at \url{https://github.com/waternumbers/ASBO/issues}
}

}
\author{
\strong{Maintainer}: Paul Smith \email{paul@waternumbers.co.uk} (\href{https://orcid.org/0000-0002-0034-3412}{ORCID})

}
\keyword{internal}
