% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_cost.R
\name{multinomialCost}
\alias{multinomialCost}
\alias{multinomialBase}
\alias{multinomialReplace}
\title{R6 class for the multinomial distribution}
\description{
Cost functions for the multinomial distribution

Compute the non-anomalous cost of a segment

Compute the segment cost for an anomalous segment with replaced rates
}
\details{
Collective anomalies are represented as chnages to the expected proportions.
Time varying expected proportions are currently not handled.
}
\examples{
set.seed(0)
m <- c(1:4)/sum(1:4)
X <- t(rmultinom(100, 144, m))

p <- multinomialBase$new(X,m,2)
p$cost(90,95) ## cost of non-anomalous distribution for x[90:95]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-multinomialCost-length}{\code{multinomialCost$length()}}
\item \href{#method-multinomialCost-new}{\code{multinomialCost$new()}}
\item \href{#method-multinomialCost-k}{\code{multinomialCost$k()}}
\item \href{#method-multinomialCost-min_length}{\code{multinomialCost$min_length()}}
\item \href{#method-multinomialCost-param}{\code{multinomialCost$param()}}
\item \href{#method-multinomialCost-clone}{\code{multinomialCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$new(x, m = rep(1/ncol(x), ncol(x)), gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{integer matrix of observations}

\item{\code{m}}{numeric vector of expected proportions}

\item{\code{gamma}}{penalty parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-k"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-k}{}}}
\subsection{Method \code{k()}}{
constant required in PELT algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$k()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-min_length"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-min_length}{}}}
\subsection{Method \code{min_length()}}{
minimum length of sample required
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$min_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ASBO:multinomialCost]{ASBO::multinomialCost}} -> \code{multinomialBase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-multinomialBase-cost}{\code{multinomialBase$cost()}}
\item \href{#method-multinomialBase-clone}{\code{multinomialBase$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="initialize"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-initialize'><code>ASBO::multinomialCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="k"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-k'><code>ASBO::multinomialCost$k()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="length"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-length'><code>ASBO::multinomialCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="min_length"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-min_length'><code>ASBO::multinomialCost$min_length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="param"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-param'><code>ASBO::multinomialCost$param()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialBase-cost"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialBase-cost}{}}}
\subsection{Method \code{cost()}}{
Compute the segment cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialBase$cost(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ASBO:multinomialCost]{ASBO::multinomialCost}} -> \code{multinomialReplace}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-multinomialReplace-cost}{\code{multinomialReplace$cost()}}
\item \href{#method-multinomialReplace-param}{\code{multinomialReplace$param()}}
\item \href{#method-multinomialReplace-clone}{\code{multinomialReplace$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="initialize"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-initialize'><code>ASBO::multinomialCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="k"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-k'><code>ASBO::multinomialCost$k()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="length"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-length'><code>ASBO::multinomialCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="multinomialCost" data-id="min_length"><a href='../../ASBO/html/multinomialCost.html#method-multinomialCost-min_length'><code>ASBO::multinomialCost$min_length()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialReplace-cost"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialReplace-cost}{}}}
\subsection{Method \code{cost()}}{
Compute the segment cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialReplace$cost(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialReplace-param"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialReplace-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialReplace$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialReplace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialReplace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialReplace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
