% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cost.R
\name{poisCost}
\alias{poisCost}
\alias{poisBase}
\alias{poisProp}
\alias{poisPropG}
\title{R6 class for Univariate Poisson Cost Functions}
\description{
Cost functions for the univariate Poission distribution

Compute the cost of a non-anomalous segment

Compute the cost of a anomalous segment with rate scaling parameter and constant cost

Compute the cost of a anomalous segment with rate scaling parameter and cost proportional to expectation
}
\details{
Collective anomalies are represented as multiplicative changes in rate
}
\examples{
set.seed(0)
r <- 8 + runif(100)*2
x <- rpois(100,lambda = r)

p <- poisProp$new(x,r,0)
p$cost(90,95) ## cost of anomalous distribution for x[90:95]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-poisCost-length}{\code{poisCost$length()}}
\item \href{#method-poisCost-new}{\code{poisCost$new()}}
\item \href{#method-poisCost-k}{\code{poisCost$k()}}
\item \href{#method-poisCost-min_length}{\code{poisCost$min_length()}}
\item \href{#method-poisCost-param}{\code{poisCost$param()}}
\item \href{#method-poisCost-clone}{\code{poisCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$new(x, rate = 1, gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector of observations}

\item{\code{rate}}{numeric vector of rate parameters}

\item{\code{gamma}}{penalty parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-k"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-k}{}}}
\subsection{Method \code{k()}}{
constant required in PELT algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$k()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-min_length"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-min_length}{}}}
\subsection{Method \code{min_length()}}{
minimum length required for evaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$min_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ASBO:poisCost]{ASBO::poisCost}} -> \code{poisBase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-poisBase-cost}{\code{poisBase$cost()}}
\item \href{#method-poisBase-clone}{\code{poisBase$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="initialize"><a href='../../ASBO/html/poisCost.html#method-poisCost-initialize'><code>ASBO::poisCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="k"><a href='../../ASBO/html/poisCost.html#method-poisCost-k'><code>ASBO::poisCost$k()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="length"><a href='../../ASBO/html/poisCost.html#method-poisCost-length'><code>ASBO::poisCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="min_length"><a href='../../ASBO/html/poisCost.html#method-poisCost-min_length'><code>ASBO::poisCost$min_length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="param"><a href='../../ASBO/html/poisCost.html#method-poisCost-param'><code>ASBO::poisCost$param()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisBase-cost"></a>}}
\if{latex}{\out{\hypertarget{method-poisBase-cost}{}}}
\subsection{Method \code{cost()}}{
Compute the segment cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisBase$cost(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-poisBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ASBO:poisCost]{ASBO::poisCost}} -> \code{poisProp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-poisProp-cost}{\code{poisProp$cost()}}
\item \href{#method-poisProp-param}{\code{poisProp$param()}}
\item \href{#method-poisProp-clone}{\code{poisProp$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="initialize"><a href='../../ASBO/html/poisCost.html#method-poisCost-initialize'><code>ASBO::poisCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="k"><a href='../../ASBO/html/poisCost.html#method-poisCost-k'><code>ASBO::poisCost$k()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="length"><a href='../../ASBO/html/poisCost.html#method-poisCost-length'><code>ASBO::poisCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="min_length"><a href='../../ASBO/html/poisCost.html#method-poisCost-min_length'><code>ASBO::poisCost$min_length()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisProp-cost"></a>}}
\if{latex}{\out{\hypertarget{method-poisProp-cost}{}}}
\subsection{Method \code{cost()}}{
Compute the segment cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisProp$cost(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisProp-param"></a>}}
\if{latex}{\out{\hypertarget{method-poisProp-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisProp$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisProp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-poisProp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisProp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ASBO:poisCost]{ASBO::poisCost}} -> \code{poisPropG}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-poisPropG-cost}{\code{poisPropG$cost()}}
\item \href{#method-poisPropG-k}{\code{poisPropG$k()}}
\item \href{#method-poisPropG-param}{\code{poisPropG$param()}}
\item \href{#method-poisPropG-clone}{\code{poisPropG$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="initialize"><a href='../../ASBO/html/poisCost.html#method-poisCost-initialize'><code>ASBO::poisCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="length"><a href='../../ASBO/html/poisCost.html#method-poisCost-length'><code>ASBO::poisCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ASBO" data-topic="poisCost" data-id="min_length"><a href='../../ASBO/html/poisCost.html#method-poisCost-min_length'><code>ASBO::poisCost$min_length()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisPropG-cost"></a>}}
\if{latex}{\out{\hypertarget{method-poisPropG-cost}{}}}
\subsection{Method \code{cost()}}{
Compute the segment cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisPropG$cost(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisPropG-k"></a>}}
\if{latex}{\out{\hypertarget{method-poisPropG-k}{}}}
\subsection{Method \code{k()}}{
constant required in PELT algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisPropG$k()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisPropG-param"></a>}}
\if{latex}{\out{\hypertarget{method-poisPropG-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisPropG$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisPropG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-poisPropG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisPropG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
